+++ HISTORY.md
@@ -12,6 +12,7 @@
 ### Added
 
 - Save the source info when the user calls the macros
+- Allow modification of the logger's alloc field
 
 ## [0.3.4] - 2024-03-13